Add ROLE_DEMO and manage demo on TagController

+++ src/Controller/Admin/TagController.php
@@ -38,11 +38,15 @@ class TagController extends AbstractController
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
-            $entityManager = $this->getDoctrine()->getManager();
-            $entityManager->persist($tag);
-            $entityManager->flush();
+            if ($this->isGranted('ROLE_ADMIN')) {
+                $entityManager = $this->getDoctrine()->getManager();
+                $entityManager->persist($tag);
+                $entityManager->flush();
 
-            $this->addFlash('success', 'Le mot clé a bien été ajouté.');
+                $this->addFlash('success', 'Le mot clé a bien été ajouté.');
+            } else {
+                $this->addFlash('warning', 'Le mot clé n\'a pas été ajouté. Vous êtes en mode Démonstration');
+            }
             return $this->redirectToRoute('app_tag_index', [], Response::HTTP_SEE_OTHER);
         }
 
@@ -64,9 +68,14 @@ class TagController extends AbstractController
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
-            $this->getDoctrine()->getManager()->flush();
+            if ($this->isGranted('ROLE_ADMIN')) {
+                $this->getDoctrine()->getManager()->flush();
 
-            $this->addFlash('success', 'Le mot clé a bien été modifié.');
+                $this->addFlash('success', 'Le mot clé a bien été modifié.');
+            } else {
+                $this->addFlash('warning', 'Le mot clé n\'a pas été modifié. Vous êtes en mode Démonstration');
+            }
+
             return $this->redirectToRoute('app_tag_index', [], Response::HTTP_SEE_OTHER);
         }
 
@@ -85,12 +94,17 @@ class TagController extends AbstractController
     public function delete(Request $request, Tag $tag): Response
     {
         if ($this->isCsrfTokenValid('delete' . $tag->getId(), $request->request->get('_token'))) {
-            $entityManager = $this->getDoctrine()->getManager();
-            $entityManager->remove($tag);
-            $entityManager->flush();
+            if ($this->isGranted('ROLE_ADMIN')) {
+                $entityManager = $this->getDoctrine()->getManager();
+                $entityManager->remove($tag);
+                $entityManager->flush();
+
+                $this->addFlash('success', 'Le mot clé a bien été supprimé.');
+            } else {
+                $this->addFlash('warning', 'Le mot clé n\'a pas été supprimé. Vous êtes en mode Démonstration');
+            }
         }
 
-        $this->addFlash('success', 'Le mot clé a bien été supprimé.');
         return $this->redirectToRoute('app_tag_index', [], Response::HTTP_SEE_OTHER);
     }
 }
